fixed reample bugs + chebeval adjustment
the chebyshev polynomial type was adjusted to contain xmin and xmax from
the inital fititng, as those need to be the same for evaluation.

for resampling, some code to average the next 2 closest values from the
position profile when a value was not withing [0,1] was added. this does
not work when 2 values outside that range are next to each other so i
will need to make this more robust in the future. the degree used to fit
arclen->curve param is slighty arbitrary, initially i planned to use
seg.rows(), but that caused overfitting, so honestly i do not know what
the best choice of degree would be.